fix: functional tests
- move validation out of language specific modules
